Add update progress
