+++ metagpt/actions/minecraft/__init__.py
@@ -7,7 +7,7 @@ from enum import Enum
 from metagpt.actions.action import Action
 from metagpt.actions.action_output import ActionOutput
 from metagpt.actions.minecraft.design_curriculumn import DesignTask, DesignCurriculum
-from metagpt.actions.minecraft.generate_actions import GenerateActionCode, SummarizeLog
+from metagpt.actions.minecraft.generate_actions import GenerateActionCode
 from metagpt.actions.minecraft.manage_skills import RetrieveSkills, GenerateSkillDescription, AddNewSkills
 from metagpt.actions.minecraft.review_task import VerifyTask
 from metagpt.actions.minecraft.player_action import PlayerActions
@@ -19,7 +19,6 @@ class ActionType(Enum):
     Design_Task = DesignTask
     Design_Curriculum = DesignCurriculum
     Generate_Action_Code = GenerateActionCode
-    Summarize_Log = SummarizeLog
     Retrieve_Skills = RetrieveSkills
     Generate_Skill_Description = GenerateSkillDescription
     Add_New_Skills = AddNewSkills
